# lib/api-zod/src/generated/types/maintenanceCollectionInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Pavilion community website API
 * OpenAPI spec version: 0.1.0
 */
import type { MaintenanceCollectionInputPaymentMode } from './maintenanceCollectionInputPaymentMode';

export interface MaintenanceCollectionInput {
  flatId: number;
  /** @minLength 1 */
  payerName: string;
  /** @minimum 1 */
  amountPaise: number;
  /** @minLength 1 */
  paymentDate: string;
  paymentMode: MaintenanceCollectionInputPaymentMode;
  /** @minLength 1 */
  forMonth: string;
  referenceNumber?: string;
  notes?: string;
}