# lib/api-zod/src/generated/types/maintenanceRateInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Pavilion community website API
 * OpenAPI spec version: 0.1.0
 */

export interface MaintenanceRateInput {
  /** @minimum 0 */
  monthlyAmountPaise: number;
}