# lib/api-zod/src/generated/types/maintenanceSettings.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Pavilion community website API
 * OpenAPI spec version: 0.1.0
 */

export interface MaintenanceSettings {
  id: number;
  dueDay: number;
  lateFeePercent: number;
  openingBalanceNote: string;
}